payments added

+++ src/main/java/com/turkcell/rentACar/business/adapters/bankAdapters/ZiraatBankAdapter.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Service;
 public class ZiraatBankAdapter implements PostService {
 
     @Override
-    public boolean addPayment(String cardOwnerName, String cardNumber, int cardCVC,int cardEndMonth, int cardEndYear,int amount) {
+    public boolean addPayment(String cardOwnerName, String cardNumber, int cardCVC,int cardEndMonth, int cardEndYear,double amount) {
 
         FakeZiraatBankManager fakeZiraatBankManager = new FakeZiraatBankManager();
         return fakeZiraatBankManager.addPayments(cardNumber,cardOwnerName,cardEndMonth,cardEndYear,cardCVC,amount);
