+++ src/main/java/com/turkcell/rentACar/core/bankServices/FakeZiraatBankManager.java
@@ -3,7 +3,7 @@ package com.turkcell.rentACar.core.bankServices;
 // Bu koda dokunulamaz...
 public class FakeZiraatBankManager {
 
-    public boolean addPayments(String cardNumber, String cardOwnerName, int cardEndMonth, int cardEndYear,int cardCVC,int amount) {
+    public boolean addPayments(String cardNumber, String cardOwnerName, int cardEndMonth, int cardEndYear,int cardCVC,double amount) {
         return true;
     }
 }
