+++ src/main/java/com/turkcell/rentACar/core/bankServices/PostService.java
@@ -2,5 +2,5 @@ package com.turkcell.rentACar.core.bankServices;
 
 public interface PostService {
 
-    boolean addPayment(String cardOwnerName, String cardNumber, int cardCVC,int cardEndMonth, int cardEndYear,int amount);
+    boolean addPayment(String cardOwnerName, String cardNumber, int cardCVC,int cardEndMonth, int cardEndYear,double amount);
 }
